Refactor RedisInstance into a separate module.
